pricefeed param set to 96bit to use full slot, ETH quote type set to 0

+++ contracts/Storage.sol
@@ -88,8 +88,9 @@ abstract contract Storage is Constants {
     }
 
     struct PriceFeedStorage {
+        // Packed slot: 20 + 12 = 32
         address priceFeed;
-        uint32 params;
+        uint96 params;
     }
 
     mapping(address => AssetConfig) internal underlyingLookup; // underlying => AssetConfig
